backend/requirements.txt
fastapi==0.116.1
numpy==2.3.3
Pillow==11.3.0
uvicorn==0.35.0